updated WRESL code to have the correct constants. update the JRXML with the latest from Bryan

+++ eppt-report-engine/src/main/java/gov/ca/water/reportengine/jython/JythonValueGenerator.java
@@ -91,7 +91,7 @@ public class JythonValueGenerator
 			throws ScriptException
 	{
 		this(epptScenarioRun, function, commonPeriodFilter);
-		_scriptRunner.setComparisonValue(comparisonValue);
+		_scriptRunner.setComparisonValue((double)comparisonValue);
 	}
 
 	public Double generateValue() throws EpptReportException
